Command Hold after mission ending in safe landing to trigger disarm
Does not automatically disarm and switch to hold in this case. Manually commanding a Hold successfully compltes the missinos and causes a disarm.
While pausing the OA trajectory message sending can have the same effect, it also puts the OA interface into a failure mode, which interferes with subsequent take-offs.